+++ options_pricer_European/utils/strategies.py
